+++ app/controllers/users/otp_challenge_controller.rb
@@ -13,13 +13,23 @@ class Users::OtpChallengeController < ApplicationController
       return
     end
 
+    if user.otp_locked?
+      clear_otp_session
+      redirect_to new_user_session_path,
+                  alert: 'Account temporarily locked due to too many failed 2FA attempts. Try again in 30 minutes or reset your password.'
+      return
+    end
+
     otp_code = params[:otp_attempt]
 
     if user.validate_and_consume_otp!(otp_code) || user.invalidate_otp_backup_code!(otp_code)
       clear_otp_session
+      user.reset_failed_otp_attempts!
       sign_in(user)
       redirect_to after_sign_in_path_for(user), notice: 'Signed in successfully.'
     else
+      user.register_failed_otp_attempt!
+
       session[:otp_failed_attempts] = (session[:otp_failed_attempts] || 0) + 1
       if session[:otp_failed_attempts] >= MAX_FAILED_ATTEMPTS
         clear_otp_session